Update-3273e3e0-12a9-457f-bfc7-08bac188f1f7

+++ src/main/java/com/bss/inc/cameraon/utils/net/angryscan/feeders/RangeFeeder.java
@@ -66,8 +66,8 @@ public class RangeFeeder extends AbstractFeeder {
         return new ScanningSubject(prevIP);
     }
 
-    public double percentageComplete() {
-        return Math.round(percentageComplete);
+    public int percentageComplete() {
+        return (int)Math.round(percentageComplete);
     }
 
     /**
